Consolidate cron tasks into QueueService
